// src/models/dc/law.ts
export interface createDC {

    uuid: string;
    name: string;
    address_1: string;
    address_2: string;
    area: string;
    city: string;
    state: string;
    pincode:string;
    latitude: string;
    longitude: string;
    status:string;
    created_on: number;
    updated_on: number;
    
}